export: use UART port 1 at high baud rate
Note that UART port 0 is the port forwarded by ICDI which probably has
a fixed baud rate of 115200. At least, we know for certain that higher
baud rates drop bytes. It is not clear whether the changing the
baud rate via stty actually is handled in ICDI (the stty does accept
the change on linux side). So, we use an external FTDI chip.